+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Unreleased
-- Bug Fix: Add **optional** fix for non compliant built-in scalar Int type. Set `config :absinthe, use_spec_compliant_int_scalar: :true` in your config to use the fixed Int type. It is also advisable to upgrade for custom types if you are leveraging the use of integers outside the standard. More information in the [pull request #1131](https://github.com/absinthe-graphql/absinthe/pull/1131).
+
+- Bug Fix: Adds **optional fix** for non compliant built-in scalar Int type. Add `@pipeline_modifier Absinthe.Phase.Schema.SpecCompliantInt` to your schema to use the fixed Int type. It is also advisable to upgrade for custom types if you are leveraging the use of integers outside the GraphQl standard.
 
 - Feature: [Convert SDL Language.\* structs to SDL notation](https://github.com/absinthe-graphql/absinthe/pull/1160)
 - Feature: [Add support for type extensions](https://github.com/absinthe-graphql/absinthe/pull/1157)